Created Page for Adding Product to Inventory Wastage
A page for adding products that are unsellable or has expired was created. A search filter feature was also added in order to efficiently search the wasted product from the main inventory and be removed. (NOTE: the data of the product isn't necessarily removed from the database, instead, it is transferred to WasteProduct to be used for data reports).